chore: Prepping for switching SlangShader to Swift
Moved historyCount to OEShaderPassCompiler, which is another step
at separating the data model for shaders and parameters from
a specific processor. OEShaderPassCompiler is specific to transforming
individual Vulkan glsl passes, which is all the framework currently
supports. The end goal is that individual passes could be written in
Vulkan glsl or Metal.