fix: Teach OEFilterChain how to clear textures that might be read
Textures used for effects such as feedback or history may contain
invalid data on first use, which may result in undefined behaviour.

An example is the Dr Venom shader, which uses feedback to determine
average luminosity. On the first use, the values in the buffer may
contain NaN values, which result all subsequent frames containing NaNs
and the output white.

It is possible this only occurs on unified hardware (such as Si).